refactor: use exists by id instead to optimize query
Took 16 minutes

+++ src/main/java/dev/realtards/kuenyawz/services/ProductServiceImpl.java
@@ -80,6 +80,10 @@ public class ProductServiceImpl implements ProductService {
 
 	@Override
 	public void hardDeleteProduct(Long productId) {
+		if (!productRepository.existsById(productId)) {
+			throw new ResourceNotFoundException("Product with ID '" + productId + "' not found");
+		}
+
 		productRepository.deleteById(productId);
 	}
